+++ src/main/java/com/girigiri/kwrental/rental/service/restore/PenaltyService.java
@@ -1,4 +1,4 @@
-package com.girigiri.kwrental.reservation.service;
+package com.girigiri.kwrental.rental.service.restore;
 
 import com.girigiri.kwrental.rental.domain.RentalSpecStatus;
 
@@ -6,8 +6,6 @@ public interface PenaltyService {
 
     void createOrUpdate(Long memberId, Long reservationId, Long reservationSpecId, Long rentalSpecId,
         RentalSpecStatus status);
-
-    boolean hasOngoingPenalty(Long memberId);
 
     void deleteByRentalSpecIdIfExists(Long rentalSpecId);
 }
