+++ src/main/java/com/girigiri/kwrental/reservation/service/PenaltyChecker.java
@@ -1,0 +1,5 @@
+package com.girigiri.kwrental.reservation.service;
+
+public interface PenaltyChecker {
+	boolean hasOngoingPenalty(Long memberId);
+}
